First commit at modularizing repo
