+++ pyproject.toml
@@ -1,0 +1,23 @@
+[tool.poetry]
+name = "caat"
+version = "0.0.1"
+description = "The Catalog of Archival Astronomical Transients"
+authors = ["Craig Pellegrino, Tyler Pritchard, Adrian Crawford"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+matplotlib = "^3.9.1"
+pandas = "^2.2.2"
+numpy = "<2.0.0"
+scikit-learn = "^1.5.1"
+scipy = "^1.14.0"
+extinction = "^0.4.6"
+astropy = "^6.1.2"
+dustmaps = "^1.0.13"
+pytest = "^8.3.2"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
