+++ src/caat/utils.py
@@ -1,0 +1,23 @@
+import os
+
+ROOT_DIR = os.path.dirname(
+    os.path.dirname(
+        os.path.dirname(
+            os.path.abspath(__file__)
+        )
+    )
+)
+
+colors = {
+    "U": "purple",
+    "B": "blue",
+    "V": "lime",
+    "g": "cyan",
+    "r": "orange",
+    "i": "red",
+    "UVW2": "#FE0683",
+    "UVM2": "#BF01BC",
+    "UVW1": "#8B06FF",
+    "c": "turquoise",
+    "o": "salmon",
+}
